rename sendNotifications.ts
